> proper project declaration > multi file script > pip installation
