+++ src/Configuration.py
@@ -1,0 +1,41 @@
+import json
+
+from src.Strategy import Strategy
+from src.exception.InvalidStrategyException import InvalidStrategyException
+
+
+class Configuration:
+    path = None
+    data = None
+
+    def __init__(self, path):
+        self.path = path
+        self.reload()
+
+    def reload(self):
+        with open(self.path, "r") as fp:
+            try:
+                self.data = json.load(fp)
+            except json.JSONDecodeError:
+                return False
+        return True
+
+    def getStrategies(self):
+        return self.data["strategies"].keys()
+
+    def getStrategy(self, strategyName):
+        if strategyName == "strategyOnDischarging":
+            strategyName = self.data[strategyName]
+            if strategyName == "":
+                strategyName = "defaultStrategy"
+        if strategyName == "defaultStrategy":
+            strategyName = self.data[strategyName]
+        if strategyName is None or strategyName not in self.data["strategies"]:
+            raise InvalidStrategyException(strategyName)
+        return Strategy(strategyName, self.data["strategies"][strategyName])
+
+    def getDefaultStrategy(self):
+        return self.getStrategy("defaultStrategy")
+
+    def getDischargingStrategy(self):
+        return self.getStrategy("strategyOnDischarging")
