+++ src/Strategy.py
@@ -1,0 +1,15 @@
+class Strategy:
+    name = None
+    fanSpeedUpdateFrequency = None
+    movingAverageInterval = None
+    speedCurve = None
+
+    def __init__(self, name, parameters):
+        self.name = name
+        self.fanSpeedUpdateFrequency = parameters["fanSpeedUpdateFrequency"]
+        if self.fanSpeedUpdateFrequency is None or self.fanSpeedUpdateFrequency == "":
+            self.fanSpeedUpdateFrequency = 5
+        self.movingAverageInterval = parameters["movingAverageInterval"]
+        if self.movingAverageInterval is None or self.movingAverageInterval == "":
+            self.movingAverageInterval = 20
+        self.speedCurve = parameters["speedCurve"]
