+++ src/exception/JSONException.py
@@ -1,0 +1,2 @@
+class JSONException(Exception):
+    pass
