+++ src/hardwareController/HardwareController.py
@@ -1,0 +1,25 @@
+from abc import ABC, abstractmethod
+
+from src.exception.UnimplementedException import UnimplementedException
+
+
+class HardwareController(ABC):
+    @abstractmethod
+    def getTemperature(self):
+        raise UnimplementedException()
+
+    @abstractmethod
+    def setSpeed(self, speed):
+        raise UnimplementedException()
+
+    @abstractmethod
+    def pause(self):
+        pass
+
+    @abstractmethod
+    def resume(self):
+        pass
+
+    @abstractmethod
+    def isOnAC(self):
+        raise UnimplementedException()
